endterm/requirements.txt
appdirs==1.4.4
asgiref==3.3.4
distlib==0.3.1
Django==3.2.3
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
filelock==3.0.12
Pillow==8.2.0
PyJWT==1.7.1
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
virtualenv==20.4.6
